Add graceful shutdown and a shutdown hook in Application.
The graceful shutdown period, during which new requests are rejected but
outstanding requests are allowed to finish, is configurable. A shutdown hook in
Application also allows additional resources not associated with any single
request to be cleaned up. In order to ensure the process does not exit before
the hook runs, lifecycle methods are added to Application that wrap and augment
the ones in Jetty's Server class.