Add RestException with additional error code and automatically convert to ErrorMessage.
Add the RestException class to represent any errors that should be automatically
converted to an ErrorMessage and set the specified HTTP status code. This is a
subclass of WebApplicationExceptionMapper so it integrates well with the rest of
Jersey. WebApplicationExceptionMapper is updated to handle this special case,
using different values for the HTTP status and error code embedded in the
response entity. This also adds a couple of subclasses for common HTTP
statuses. Fixes #3.