Upgrade Jetty, Jersey, Jackson and JUnit to the latest released version
Jetty 9.2 includes a number of incompatible changes that we had to adapt
to. Behaviour should be the same apart from `connection-accepted-rate`
metric. `NetworkTrafficListener` doesn't expose the information we need,
so we approximate it via `opened` and deprecate it.

Use `JettyTestContainer` and `JettyTestContainerFactory` from Jersey and
delete our copy (which needed some changes to work with Jetty 9.2).

Fixes #19.