KREST-1030 Upgrade Jetty to fix a number of CVEs.
This brings the https://github.com/confluentinc/rest-utils/pull/249
and the https://github.com/confluentinc/rest-utils/pull/151 changes
to the older branches.

In addition to that, it also makes the smallest possible changes in
order to fix failures in SslTest.java after the upgrade. This doesn't
cleanly map to an existing change in the higher version branches,
because the problem was addressed before a Jetty upgrade started
causing SslTest.java to fail, but the closest equivalent is probably
the combination of https://github.com/confluentinc/rest-utils/pull/147
and https://github.com/confluentinc/rest-utils/pull/155.