KREST-661: Add support for HTTP/2 on Java 11 and later (#258)
Added support for HTTP/2 on Java 11 and later. Because HTTP/2 and ALPN are newer than Java 8, adding support for Java 8 as well is not practical. So, the behaviour is:

```
if java 11 compatible
  if (http)
    configure HTTP and HTTP/2 Clear text connection factories
  else (https)
    configure SSL, ALPN, HTTP/2 and HTTP connection factories
else as before
```

This means that when the code is running on a Java 11 JVM or later, connections will default to HTTP/2 not HTTP/1.1 if the client can support it. This change in behaviour is appropriate for the major version bump to CP 7.0 but not CP 6.x.

Added a new configuration `http2.enabled` which defaults to true. This gives a way of turning HTTP/2 off if it proves troublesome in some environment.

The existing unit tests use HTTP/1.1 and continue to be valid. A new test has been added to try the various combinations of HTTP/1.1, HTTP/2 and configuration.