MMA-12033 Bump timeouts for connections limits testing
The connection limits functionality introduced with KREST-4977
came with a test that sends requests and waits up to some small
timeout for a response. In the negative case, when connections
should be limited, the test expects the request to fail after
<timeout> seconds, and in the positive case, when connections
should be reenabled, the test expects the request to be handled
and a response to be received within <timeout> seconds.

Despite the test succeeding 100 out of 100 times locally, it
fails with impressive consistency on Jenkins. One reason for
that could be the small 1-second timeout which might not be
enough on Jenkins. Another might be that space for only a
single connection is being made before asserting the positive
case.