KNET-16654: Handle jetty 11+ behavior of NetworkTrafficServerConnector (#516)
There is a breaking change in NetworkTrafficServerConnector class from accepting multiple NetworkTrafficListener (jetty 9) to only accept a single listener (jetty 11+)

We currently have two types of listeners, one for handling metrics and one for rate limiting. We can merge the two listeners into a single class with shared logic from both. Then we register the single listener into the NetworkTrafficServerConnector

We will also need to change the API call from addNetworkTrafficListener to setNetworkTrafficListener when we bump the jetty version